// src/pages/calculator/Calculator.enums.ts
export enum CalculatorFieldsEnum {
	Material = 'material',
	Pipe = 'pipe',
	Width = 'width',
	Length = 'length',
	Frame = 'frame',
}

export enum CalculatorDataType {
	List = 'list',
	Pipe = 'pipe',
	Fix = 'fix',
}
export enum CalculatorConfigType {
	Size = 'size',
	Material = 'material',
	Fix = 'fix',
	Frame = 'frame',
}

export enum CalculatorDataUnit {
	M2 = 'м2',
	MP = 'мп',
	Piece = 'шт',
}

export enum FieldTypesEnum {
	Dropdown = 'dropdown',
	Number = 'number',
}